Repackage contact- and event-related commands
