+++ src/main/java/seedu/address/logic/commands/contacts/ClearContactCommand.java
@@ -1,7 +1,11 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.contacts;
 
 import static java.util.Objects.requireNonNull;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.CommandType;
+import seedu.address.logic.commands.CommandWord;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
