+++ src/main/java/seedu/address/logic/commands/contacts/PermaSortContactCommand.java
@@ -1,40 +1,45 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.contacts;
 
-import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.CommandType;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.comparators.PersonComparator;
 import seedu.address.model.Model;
 
+
 /**
- * Sorts the currently displayed persons in a specific order.
+ * Sorts the contacts in Athena's address book permanently.
  * Index entered determines the specific order.
  */
-public class SortContactCommand extends Command {
+public class PermaSortContactCommand extends Command {
 
-    public static final String COMMAND_WORD = CommandWord.SORT.toString();
+    public static final String COMMAND_WORD = "psort";
 
     public static final String COMMAND_TYPE = CommandType.CONTACT.toString();
 
     public static final String MESSAGE_USAGE = COMMAND_WORD + " " + COMMAND_TYPE
-            + ": Sorts the currently displayed list"
-            + "by the index command entered "
+            + ": Sort your contacts permanently "
+            + "by the index command entered  "
             + "1 will be sort by alphabetical order of their names\n"
             + "2 will be sort by alphabetical order of their address\n"
             + "3 will be sort by alphabetical order of their email\n"
-            + "Parameters: INDEX (must be between 1 and 4) "
+            + "Parameters: INDEX (must be between 1 and 3) "
             + "Example: " + COMMAND_WORD + " " + COMMAND_TYPE + " 1 ";
 
-    public static final String MESSAGE_ARGUMENTS = "Index: %1$d";
+    public static final String MESSAGE_INVALID_INDEX = "Invalid index entered, "
+            + "refer to below for the command's proper usage: "
+            + MESSAGE_USAGE;
 
     private final Index index;
 
     /**
      * @param index the order in which to sort the address book
      */
-    public SortContactCommand(Index index) {
+    public PermaSortContactCommand(Index index) {
         requireAllNonNull(index);
         this.index = index;
     }
@@ -42,10 +47,17 @@ public class SortContactCommand extends Command {
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
-        model.sortPerson(PersonComparator.chooseComparator(index));
+        if (index.getOneBased() < 0 || index.getOneBased() > 3) {
+            throw new CommandException(PermaSortContactCommand.MESSAGE_INVALID_INDEX);
+        }
+        model.permaSortContacts(PersonComparator.chooseComparator(index));
 
         return new CommandResult(indexMessage(index));
     }
+
+    private void requireNonNull(Model model) {
+    }
+
 
     /**
      * Returns the appropriate console message for the index.
@@ -53,6 +65,7 @@ public class SortContactCommand extends Command {
     public String indexMessage(Index index) {
         int input = index.getOneBased();
         switch (input) {
+
         case 1:
             return "Sorted by name in alphabetical order";
         case 2:
@@ -60,8 +73,7 @@ public class SortContactCommand extends Command {
         case 3:
             return "Sorted by email in alphabetical order";
         default:
-            return "Invalid index entered, refer to below for the command's proper usage: "
-                    + MESSAGE_USAGE;
+            return MESSAGE_INVALID_INDEX;
         }
     }
 
@@ -73,13 +85,12 @@ public class SortContactCommand extends Command {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof SortContactCommand)) {
+        if (!(other instanceof PermaSortContactCommand)) {
             return false;
         }
 
         // state check
-        SortContactCommand e = (SortContactCommand) other;
+        PermaSortContactCommand e = (PermaSortContactCommand) other;
         return index.equals(e.index);
     }
-
 }
