+++ src/main/java/seedu/address/logic/commands/events/ClearEventCommand.java
@@ -1,7 +1,11 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.events;
 
 import static java.util.Objects.requireNonNull;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.CommandType;
+import seedu.address.logic.commands.CommandWord;
 import seedu.address.model.Calendar;
 import seedu.address.model.Model;
