+++ src/main/java/seedu/address/logic/commands/events/FindEventCommand.java
@@ -1,8 +1,12 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.events;
 
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.Messages;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.CommandType;
+import seedu.address.logic.commands.CommandWord;
 import seedu.address.model.Model;
 import seedu.address.model.event.DescriptionContainsKeywordsPredicate;
