+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -6,16 +6,16 @@ import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import seedu.address.logic.commands.ClearContactCommand;
-import seedu.address.logic.commands.ClearEventCommand;
+import seedu.address.logic.commands.contacts.ClearContactCommand;
+import seedu.address.logic.commands.events.ClearEventCommand;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandType;
 import seedu.address.logic.commands.CommandWord;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.IntroCommand;
-import seedu.address.logic.commands.ListContactCommand;
-import seedu.address.logic.commands.ListEventCommand;
+import seedu.address.logic.commands.contacts.ListContactCommand;
+import seedu.address.logic.commands.events.ListEventCommand;
 import seedu.address.logic.commands.ListTagCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
