+++ src/main/java/seedu/address/logic/parser/PermaSortContactCommandParser.java
@@ -4,7 +4,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.logic.commands.PermaSortContactCommand;
+import seedu.address.logic.commands.contacts.PermaSortContactCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
