+++ src/main/java/seedu/address/logic/parser/SortContactCommandParser.java
@@ -4,7 +4,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.logic.commands.SortContactCommand;
+import seedu.address.logic.commands.contacts.SortContactCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
