+++ src/main/java/seedu/address/logic/parser/SortEventCommandParser.java
@@ -4,7 +4,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.logic.commands.SortEventCommand;
+import seedu.address.logic.commands.events.SortEventCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
