+++ src/test/java/seedu/address/logic/commands/FindContactCommandTest.java
@@ -16,6 +16,7 @@ import java.util.Collections;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.contacts.FindContactCommand;
 import seedu.address.logic.parser.FindContactCommandParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Calendar;
