+++ src/test/java/seedu/address/logic/commands/PermaSortCommandTest.java
@@ -9,6 +9,7 @@ import java.util.Comparator;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.contacts.PermaSortContactCommand;
 import seedu.address.model.Calendar;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
