+++ src/test/java/seedu/address/logic/parser/AddContactCommandParserTest.java
@@ -27,7 +27,7 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.AddContactCommand;
+import seedu.address.logic.commands.contacts.AddContactCommand;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
