+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -17,21 +17,21 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.AddContactCommand;
-import seedu.address.logic.commands.AddEventCommand;
-import seedu.address.logic.commands.ClearContactCommand;
-import seedu.address.logic.commands.DeleteContactCommand;
-import seedu.address.logic.commands.DeleteEventCommand;
-import seedu.address.logic.commands.EditContactCommand;
-import seedu.address.logic.commands.EditContactCommand.EditPersonDescriptor;
-import seedu.address.logic.commands.EditEventCommand;
-import seedu.address.logic.commands.EditEventCommand.EditEventDescriptor;
+import seedu.address.logic.commands.contacts.AddContactCommand;
+import seedu.address.logic.commands.events.AddEventCommand;
+import seedu.address.logic.commands.contacts.ClearContactCommand;
+import seedu.address.logic.commands.contacts.DeleteContactCommand;
+import seedu.address.logic.commands.events.DeleteEventCommand;
+import seedu.address.logic.commands.contacts.EditContactCommand;
+import seedu.address.logic.commands.contacts.EditContactCommand.EditPersonDescriptor;
+import seedu.address.logic.commands.events.EditEventCommand;
+import seedu.address.logic.commands.events.EditEventCommand.EditEventDescriptor;
 import seedu.address.logic.commands.ExitCommand;
-import seedu.address.logic.commands.FindContactCommand;
-import seedu.address.logic.commands.FindEventCommand;
+import seedu.address.logic.commands.contacts.FindContactCommand;
+import seedu.address.logic.commands.events.FindEventCommand;
 import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.ListContactCommand;
-import seedu.address.logic.commands.ListEventCommand;
+import seedu.address.logic.commands.contacts.ListContactCommand;
+import seedu.address.logic.commands.events.ListEventCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.event.DescriptionContainsKeywordsPredicate;
 import seedu.address.model.event.Event;
