+++ src/test/java/seedu/address/logic/parser/FindContactCommandParserTest.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.FindContactCommand;
+import seedu.address.logic.commands.contacts.FindContactCommand;
 import seedu.address.model.person.ContactContainsFieldsPredicate;
 import seedu.address.model.tag.Tag;
